Avoid boilerplate code with a big universal delegation header (#3294)
To get an idea, this is a sample from the header:

switch (DELEGATE_ID) {
[..]
  case Expression::Id::BlockId: {
    DELEGATE_START(Block);
    DELEGATE_FIELD_CHILD_LIST(Block, list);
    DELEGATE_FIELD_SCOPE_NAME(Block, name);
    DELEGATE_END(Block);
    break;
  }

A switch is defined (on some DELEGATE_ID that is provided). There is then
a case for each expression type, calling a delegate for each field type. This allows
a user to define the delegates it wants, then include this header.

This uses the new header in the copying logic, where the rewrite also
avoids recursion.

This fixes a missing BrOnExn::sent scanning in the existing
hashing/comparing code (which is what I used as a basis). When this is
used there, it will fix that bug.