OptimizeInstructions: Fix regression from #3303 / #3275 (#3338)
X - Y <= 0

=>

X <= Y

That is true mathematically, but not in the case of an overflow, e.g.
X=10, Y=0x8000000000000000. X - Y is a negative number, so
X - Y <= 0 is true. But it is not true that X <= Y (as Y is negative, but
X is not).

See discussion in #3303 (comment)

The actual regression was in #3275, but the fuzzer had an easier time
finding it due to #3303