[Fuzzer] Add a chance to pick particularly important initial contents (#3343)
OptimizeInstructions is seeing the most work these days, so it's good for
the fuzzer to focus on that some more.

Also move some code around in the main test wast: it's useful to put each
feature in its own module to maximize the chance of getting them to be used.
That is, if a module has a single use of atomics, then if atomics are disabled
in the current run, we can't use any of the module and we skip initial contents
entirely. Moving each feature to it's own module reduces that risk. (We do
pick randomly between the modules, and atm a small module has the same
chance as a big one, but this still seems worth it.)