Module splitting (#3317)
Adds the capability to programatically split a module into a primary and
secondary module such that the primary module can be compiled and run before the
secondary module has been instantiated. All calls to secondary functions (i.e.
functions that have been split out into the secondary module) in the primary
module are rewritten to be indirect calls through the table. Initially, the
table slots of all secondary functions contain references to imported
placeholder functions. When the secondary module is instantiated, it will
automatically patch the table to insert references to the original functions.

The process of module splitting involves these steps:

 1. Create the new secondary module.

 2. Export globals, events, tables, and memories from the primary module and
    import them in the secondary module.

 3. Move the deferred functions from the primary to the secondary module.

 4. For any secondary function exported from the primary module, export in
    its place a trampoline function that makes an indirect call to its
    placeholder function (and eventually to the original secondary function),
    allocating a new table slot for the placeholder if necessary.

 5. Rewrite direct calls from primary functions to secondary functions to be
    indirect calls to their placeholder functions (and eventually to their
    original secondary functions), allocating new table slots for the
    placeholders if necessary.

 6. For each primary function directly called from a secondary function, export
    the primary function if it is not already exported and import it into the
    secondary module.

 7. Replace all references to secondary functions in the primary module's table
    segments with references to imported placeholder functions.

 8. Create new active table segments in the secondary module that will replace
    all the placeholder function references in the table with references to
    their corresponding secondary functions upon instantiation.

Functions can be used or referenced three ways in a WebAssembly module: they can
be exported, called, or placed in a table. The above procedure introduces a
layer of indirection to each of those mechanisms that removes all references to
secondary functions from the primary module but restores the original program's
semantics once the secondary module is instantiated. As more mechanisms that
reference functions are added in the future, such as ref.func instructions, they
will have to be modified to use a similar layer of indirection.

The code as currently written makes a few assumptions about the module that is
being split:

 1. It assumes that mutable-globals is allowed. This could be worked around by
    introducing wrapper functions for globals and rewriting secondary code that
    accesses them, but now that mutable-globals is shipped on all browsers,
    hopefully that extra complexity won't be necessary.

 2. It assumes that all table segment offsets are constants. This simplifies the
    generation of segments to actively patch in the secondary functions without
    overwriting any other table slots. This assumption could be relaxed by 1)
    having secondary segments re-write primary function slots as well, 2)
    allowing addition in segment offsets, or 3) synthesizing a start function to
    modify the table instead of using segments.

 3. It assumes that each function appears in the table at most once. This isn't
    necessarily true in general or even for LLVM output after function
    deduplication. Relaxing this assumption would just require slightly more
    complex code, so it is a good candidate for a follow up PR.

Future Binaryen work for this feature includes providing a command line tool
exposing this functionality as well as C API, JS API, and fuzzer support. We
will also want to provide a simple instrumentation pass for finding dead or
late-executing functions that would be good candidates for splitting out. It
would also be good to integrate that instrumentation with future function
outlining work so that dead or exceptional basic blocks could be split out into
a separate module.