[effects.h] Make internals internal, forcing the external API to be safe. (#3385)
A user of EffectAnalyzer could call walk or visit, to walk the entire
input or just visit the node without chlidren. But this was unsafe - we just
exposed the Walker/Visitor API here, and did not ensure that users did
the stuff in analyze which does a little stuff before and after. In fact
Vacuum got this wrong.

To avoid that, move all the internals to an internal class. The external
API now only allows the caller to call walk or visit, and both are safe.

The change here is mostly whitespace + adding parent. prefixes.

This is NFC except for fixing possible Vacuum issues (which I am not sure
could happen in practice or not).