[wasm-type] Introduce a function for walking type trees
Since types can be arbitrarily deeply nested, even discovering all the types
used in a module requires being able to traverse type definition trees. This PR
introduces a function to do just that. It takes two callbacks, a function to
call on each visited type and a function to call whenever a recursion is
encountered. Note that it is not currently possible to construct a recursive
type, but the recursion callback will be useful in an upcoming PR to allow
constructing recursive types.