[GC types] Refactoring to allow future heap type parsing. NFC (#3409)
Defined types in wasm are really one of the "heap types": a signature type, or
(with GC) a struct or an array type. This refactors the binary and text parsers
to load the defined types into an array of heap types, so that we can start to
parse GC types. This replaces the existing array of signature types (which
could not support a struct or an array).

Locally this PR can parse and print as text simple GC types. For that it was
necessary to also fix Type::getFeatures for GC.