wasm-emscripten-finalize: Support PIC + threads
With PIC + threads the offset of passive segments is not constant but
relative to `__memory_base`.

When trying to find passive segment offset based on the target of the
`memory.init` instruction we need to consider this possibility as well as
the regular constant one.

For the llvm side of this that generates the calls to memory.init
see:  https://reviews.llvm.org/D92620