[GC] Add struct.new and start to test interesting execution (#3433)
With struct.new read/write support, we can start to do interesting
things! This adds a test of creating a struct and seeing that references
behave like references, that is, if we write to the value X refers to, and
if Y refers to the same thing, when reading from Y's value we see the
change as well.

The test is run through all of -O1, which uncovered a minor issue in
Precompute: We can't try to precompute a reference type, as we can't
replace a reference with a value.

Note btw that the test shows the optimizer properly running
CoalesceLocals on reference types, merging two locals.