[GC] Fix Array optimization issues (#3438)
Precompute still tried to precompute a reference because the check was
not in the topmost place.

Also we truncated i8/i16 values, but did not extend them properly. That
was also an issue with structs.

The new test replaces the old one by moving from -O1 to -Oz (which
runs more opts, and would have noticed this earlier), and adds array
operations too, including sign extends.