[DWARF] Correctly update .debug_addr DWARF section.
This is a minimal patch to add support for also updating the addresses
stored in the .debug_addr DWARF section, which is used to perform
address resolution at debug time with debug fission. This is minimal in
the sence that it only supports the pre-DWARFv5 .debug_addr format that
is emitted by Emscripten currently, and which only consists of a simple
list of addresses.

Ideally the copy of DWARFYAML in the Binaryen tree should be updated
with LLVM ToT at some point, which will provide full support for DWARFv5
index tables, including the new .debug_addr format.

Ref: WebAssembly/binaryen#3460
Ref: emscripten-core/emscripten#13099
Bug: https://crbug.com/1161422