Made FuncCastEmulation make and export thunks for exports
Currently it only outputs thunks for things in the wasm table.
This breaks dynamic linking emscripten SIDE_MODULES because they don't have a table.
This means dlopen and dlsym don't work if EMULATE_FUNCTION_POINTER_CASTS==1,
because EMULATE_FUNCTION_POINTER_CASTS in the calling module makes all calls
have the fixed call signature, whereas the returns from dlsym are still the original
call signature. In combination with a version of emscripten which knows about this
change (separate pull request headed their way), function pointers work through
dlsym, as does passing a function pointer out of a module, which currently breaks
as the function pointer in the wasm table is still to the old function, whereas
with this fix emscripten can hook it up to the fixed call signature version..