Remove extra-unreachable.wast
This test seems to be added in #2266 to test custom unreachable
generation in `BinaryenIRWriter`, but given that the `fromBinary` files
only contain a single `unreaachable` for the whole function, I don't
think this test serves a lot of purpose. Also the custom unreachable
generation logic in #2266 was largely replaced in #3110.